refactor(web): drop the dismiss control from the bell's briefing recipe
The notifications bell empty state offered a briefing recipe with a small
X pinned to the card's corner, plus the stored timestamp that made the
dismissal stick. The control never earned its place: it straddled the
card's edge to clear the trailing arrow, so it read as loose on the panel
rather than fixed to anything, and it put a second target next to the one
thing the scene is asking for.

The card is already gated on the user having no schedule of their own, so
anyone who acts on the offer stops seeing it, and everyone else sees the
icon well and the title alone. That gate is the dismissal.

Gone with the X: the `onDismissBriefingRecipe` prop, the
`vellum:notificationsBell:briefingRecipeDismissedAt` store and its
private-window readability probe, and the `dismissBriefingRecipe` string
in the en, es, and ru catalogs. The hook is left answering one question,
so it is now `useShouldOfferBriefingRecipe` and returns a boolean.
