+++ clients/web/src/domains/home/components/notifications-bell-empty-state.tsx
@@ -1,4 +1,4 @@
-import { Bell, Sparkles, X } from "lucide-react";
+import { Bell, Sparkles } from "lucide-react";
 import { useNavigate } from "react-router";
 
 import {
@@ -9,19 +9,15 @@ import {
 } from "@/components/empty-state-scene";
 import { useTranslation } from "@/i18n";
 import { navigateToNewConversation } from "@/utils/conversation-navigation";
-import { Button } from "@vellumai/design-library";
 
 export interface NotificationsBellEmptyStateProps {
   /** Dismisses the panel before the recipe navigates away from it. */
   onLaunchRecipe?: () => void;
   /**
    * Whether to offer the briefing recipe. The caller decides: the card is an
-   * advertisement for schedules, so it is shown only to people who have none
-   * and have not waved it away.
+   * advertisement for schedules, so it is shown only to people who have none.
    */
   showBriefingRecipe?: boolean;
-  /** Hides the recipe for good. Omit to render the card without a dismiss. */
-  onDismissBriefingRecipe?: () => void;
 }
 
 /**
@@ -35,10 +31,9 @@ export interface NotificationsBellEmptyStateProps {
  *
  * Under that it may offer the schedule that would fill the panel, seeding a
  * fresh conversation with the prompt that asks the assistant to build it. That
- * offer is aimed at people who have not adopted schedules, so `showBriefingRecipe`
- * gates it and a dismiss control retires it permanently. With the card gone
- * the icon well and the title stand alone, which is the whole scene an
- * established user ever needs.
+ * offer is aimed at people who have not adopted schedules, so
+ * `showBriefingRecipe` gates it, and anyone who has one gets the icon well and
+ * the title alone, which is the whole scene an established user ever needs.
  *
  * Cut to what fits a popover: no description and no secondary action. The
  * panel already names itself "Notifications" above this, so a description
@@ -51,7 +46,6 @@ export interface NotificationsBellEmptyStateProps {
 export function NotificationsBellEmptyState({
   onLaunchRecipe,
   showBriefingRecipe = false,
-  onDismissBriefingRecipe,
 }: NotificationsBellEmptyStateProps) {
   const { t } = useTranslation("home");
   const navigate = useNavigate();
@@ -70,38 +64,15 @@ export function NotificationsBellEmptyState({
       recipes={
         showBriefingRecipe ? (
           <EmptyStateRecipeGrid>
-            {/*
-             * The dismiss sits beside the card rather than inside it: the
-             * recipe is itself a button, and a button cannot nest one. It
-             * straddles the top-right corner, which keeps it clear of the
-             * card's trailing arrow, and carries the card's own border and
-             * surface so it reads as fixed to that corner rather than
-             * floating loose on the panel. Always present rather than
-             * revealed on hover, so keyboard and touch reach it too.
-             */}
-            <div className="relative">
-              <EmptyStateRecipeCard
-                icon={Sparkles}
-                title={t("notificationsBellEmptyState.briefingRecipeTitle")}
-                meta={t("notificationsBellEmptyState.briefingRecipeMeta")}
-                description={t(
-                  "notificationsBellEmptyState.briefingRecipeDescription",
-                )}
-                onSelect={handleSelectBriefing}
-              />
-              {onDismissBriefingRecipe ? (
-                <Button
-                  variant="ghost"
-                  size="compact"
-                  className="absolute -right-1.5 -top-1.5 z-10 h-5 w-5 rounded-full border border-[var(--border-base)] bg-[var(--surface-lift)] text-[var(--content-tertiary)] hover:bg-[var(--surface-hover)] hover:text-[var(--content-default)]"
-                  iconOnly={<X size={12} />}
-                  aria-label={t(
-                    "notificationsBellEmptyState.dismissBriefingRecipe",
-                  )}
-                  onClick={onDismissBriefingRecipe}
-                />
-              ) : null}
-            </div>
+            <EmptyStateRecipeCard
+              icon={Sparkles}
+              title={t("notificationsBellEmptyState.briefingRecipeTitle")}
+              meta={t("notificationsBellEmptyState.briefingRecipeMeta")}
+              description={t(
+                "notificationsBellEmptyState.briefingRecipeDescription",
+              )}
+              onSelect={handleSelectBriefing}
+            />
           </EmptyStateRecipeGrid>
         ) : undefined
       }
