+++ clients/web/src/domains/home/components/notifications-bell.test.tsx
@@ -291,9 +291,6 @@ const READ_LABEL = 'aria-label="Notifications"';
 /** The recipe card, matched by the title it leads with. */
 const RECIPE_LABEL = /^Set up a morning briefing/;
 
-/** The control that retires the recipe, matched by its accessible name. */
-const RECIPE_DISMISS_LABEL = "Dismiss the morning briefing suggestion";
-
 const THREE_HOURS_MS = 3 * 60 * 60 * 1000;
 
 function bellItem(overrides: Partial<FeedItem>): FeedItem {
@@ -370,8 +367,6 @@ beforeEach(() => {
   schedulesEnabledCalls.length = 0;
   skillsEnabledCalls.length = 0;
   activeAssistantIdRef.value = "assistant-1";
-  // The recipe's dismissal is the one piece of state that outlives a render,
-  // so each test starts from a store that has never been written to.
   localStorage.clear();
   updateStatusCalls.length = 0;
   triggerActionCalls.length = 0;
@@ -602,15 +597,12 @@ describe("NotificationsBell empty state", () => {
   });
 
   test("carries no second call to action beside the recipe", async () => {
-    // The bell trigger, the recipe, and the control that retires it. Nothing
-    // else: a second call to action would compete with the one thing this
-    // scene is asking for, and the dismiss is not one, it is the way out.
+    // The bell trigger and the recipe. Nothing else: a second call to action
+    // would compete with the one thing this scene is asking for.
     await openBell();
 
-    expect(screen.getAllByRole("button")).toHaveLength(3);
-    expect(
-      screen.getByRole("button", { name: RECIPE_DISMISS_LABEL }),
-    ).toBeTruthy();
+    expect(screen.getAllByRole("button")).toHaveLength(2);
+    expect(screen.getByRole("button", { name: RECIPE_LABEL })).toBeTruthy();
   });
 
   test("the recipe closes the panel and seeds a conversation", async () => {
@@ -725,30 +717,6 @@ describe("NotificationsBell briefing recipe gating", () => {
     await openBell();
 
     expect(screen.queryByRole("button", { name: RECIPE_LABEL })).toBeNull();
-  });
-
-  test("dismissing the recipe retires it", async () => {
-    await openBell();
-
-    fireEvent.click(screen.getByRole("button", { name: RECIPE_DISMISS_LABEL }));
-    await act(async () => {});
-
-    expect(screen.queryByRole("button", { name: RECIPE_LABEL })).toBeNull();
-    expect(screen.getByText("Nothing yet.")).toBeTruthy();
-  });
-
-  test("the dismissal survives a remount", async () => {
-    await openBell();
-    fireEvent.click(screen.getByRole("button", { name: RECIPE_DISMISS_LABEL }));
-    await act(async () => {});
-
-    cleanup();
-    await openBell();
-
-    expect(screen.queryByRole("button", { name: RECIPE_LABEL })).toBeNull();
-    expect(
-      screen.queryByRole("button", { name: RECIPE_DISMISS_LABEL }),
-    ).toBeNull();
   });
 
   test("hides the recipe before the active assistant resolves", async () => {
