+++ clients/web/src/i18n/locales/ru/home.json
@@ -68,7 +68,6 @@
     "briefingRecipeTitle": "Настроить утреннюю сводку",
     "briefingRecipeMeta": "Ежедневно",
     "briefingRecipeDescription": "Первое уведомление придёт завтра в 8:00.",
-    "briefingRecipePrompt": "Настрой мне утреннюю сводку: каждый будний день в 8:00 проверяй мой календарь, входящие и новости, а затем публикуй краткую сводку в моих уведомлениях.",
-    "dismissBriefingRecipe": "Скрыть предложение об утренней сводке"
+    "briefingRecipePrompt": "Настрой мне утреннюю сводку: каждый будний день в 8:00 проверяй мой календарь, входящие и новости, а затем публикуй краткую сводку в моих уведомлениях."
   }
 }
